Add Asteroid Spawner level
Removed automatic asteroid spawning randomly and instead created an "AsteroidSpawner"
type that spawns asteroids at a predictable rate. Created an example level (5) that
creates a narrow-ish window you need to shoot through in order to avoid asteroids

Added support for losing when you hit an asteroid and restarting the level. This involved
refactoring how levels are defined so we can regenerate them on demand.